decouple stuff a bit more
